+++ high_level/python/examples/e3_auto_state_switch.py
@@ -13,6 +13,7 @@ STATES = [
     "walk",
     "rl",
     "flying_trot",
+    "choreo",
     "wave",
     "dance0",
     "jump",